+++ src/virtual_client.cpp
@@ -14,8 +14,8 @@
 #include <modules/audio_device/include/audio_device.h>
 #include <modules/audio_device/include/audio_device_factory.h>
 #include <modules/audio_processing/include/audio_processing.h>
-#include <rtc_base/logging.h>
 #include <rtc_base/crypto_random.h>
+#include <rtc_base/logging.h>
 
 std::shared_ptr<VirtualClient> VirtualClient::Create(
     VirtualClientConfig config) {